% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randquote.R
\name{randquote_simple}
\alias{randquote_simple}
\title{Minimal version of getting random quotes (only quotes)}
\usage{
randquote_simple(decoded = T)
}
\arguments{
\item{decoded}{TRUE if the output text is html decoded or not (defaut TRUE)}
}
\value{
A string of quote along with the author name
}
\description{
Minimal version of getting random quotes (only quotes)
}
\examples{
randquote_simple()
}
