% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randquotes-package.R
\docType{package}
\name{randquotes}
\alias{randquotes}
\alias{randquotes-package}
\title{Get Random Quotes from Quotes on Design API}
\description{
This package connects to the site \url{http://quotesondesign.com/}
that uses the WordPress JSON REST API to provide
a way for you to grab quotes.
}
\details{
...
}
\section{See Also}{

Useful links:
\itemize{
\item \url{http://quotesondesign.com/}
\item \url{https://github.com/amrrs/randquotes}
\item Report bugs at \url{https://github.com/amrrs/randquotes/issues}
}
}

\author{
AbdulMajedRaja RS(amrrs.data@gmail.com)
}
