% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unescape_html.R
\name{unescape_html}
\alias{unescape_html}
\title{Tidy up HTML text string}
\usage{
unescape_html(str)
}
\arguments{
\item{char_string}{a character string containg HTML text with tags, page breaks, etc.}
}
\value{
a tidy character string
}
\description{
Tidy up HTML text string
}
\examples{
quote <- randquote_simple()
cat(unescape_html(quote))
}
